cmsdk: Add DeviceKeyHandler overlays
As per change Iaaab737f1501a97d7016d8d519ccf127ca059218 in fw/b,
we're now allowing devices to specify an array of DeviceKeyHandlers.

Add the appropriate array overlay for this.

Also, specify CMParts as a DeviceKeyHandler, since almost all devices
with touchscreen gestures will have to use it. The only exceptions
are the devices which register touchscreen gestures through the SensorManager.

Change-Id: Ie8e4901712d03ce4b2d799cf0ee5fc2e1bad6aae
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>