PerformanceManager: Don't force powersave mode with powersave profile
Allow to use the powersave profile without enabling the powersave
mode, but force the powersave profile when enabling the powersave
mode. This makes the profile a bit more useful since unlike the
the powersave mode it has no direct UX impact (e.g. animations),
yet it allows to reduce the power consumption.

Change-Id: Ie9312a51d58950b15cf54fb1cd0c2e1d7bcd36b6
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>